Meal type id long (#61)
* id is long in MealTypeRepo
* id is long in domain tier
* id is long in web tier

+++ domain-tier/src/main/java/com/outdoor/foodcalc/domain/model/meal/MealRef.java
@@ -29,7 +29,7 @@ public class MealRef implements IValueObject<MealRef>, FoodDetails, ProductsCont
         return meal.getMealId();
     }
 
-    public int getTypeId() {
+    public long getTypeId() {
         return meal.getType().getTypeId();
     }
